Optimize Zac Efron

+++ catscripts/Entertainment/Actors_and_actresses/Zac_Efron/ZacEfron.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 
-export KEYWORDS_ACTOR_ZACEFRON="Zac(| )Efron"
+KEYWORDS_ACTOR_ZACEFRON="Zac(| )Efron"
 
-egrep -i "$KEYWORDS_ACTOR_ZACEFRON" newpages.txt >> ZacEfron.txt
-
-ZACEFRON=`stat --print=%s ZacEfron.txt`
-
-if [ $ZACEFRON -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="ZacEfron.txt"
-  export CATNAME="Zac Efron"
-  $CATEGORIZE
-fi
 
-rm ZacEfron.txt
+  ZACEFRON=`egrep -i "$KEYWORDS_ACTOR_ZACEFRON" newpages.txt`
+
+  if [ "$ZACEFRON" != "" ];
+  then
+    egrep -i "$KEYWORDS_ACTOR_ZACEFRON" newpages.txt > ZacEfron.txt
+    export CATFILE="ZacEfron.txt"
+    export CATNAME="Zac Efron"
+    $CATEGORIZE
+    rm ZacEfron.txt
+    unset ZACEFRON
+  fi
+
+fi
